refactor: 숫자 야구 게임 코어 부분 클래스 분리
